refactor(loader): move EP tests into ep-* spec files
- ep-s3-cache.spec.ts: replaces s3-util.spec.ts (entirely new file)
- ep-loader-fast-path.spec.ts: fast-path tests extracted from gen-code-bundle.spec.ts
- gen-code-bundle.spec.ts: restored to upstream version (no EP additions)
